+++ recipes/avahi/all/conanfile.py
@@ -45,8 +45,8 @@ class AvahiConan(ConanFile):
         basic_layout(self, src_folder="src")
 
     def requirements(self):
-        self.requires("glib/2.78.1")
-        self.requires("expat/2.5.0")
+        self.requires("glib/2.78.3")
+        self.requires("expat/[>=2.6.2 <3]")
         self.requires("libdaemon/0.14")
         self.requires("dbus/1.15.8")
         self.requires("gdbm/1.23")