+++ recipes/freeglut/all/test_package_module/conanfile.py
@@ -14,7 +14,6 @@ class TestPackageConan(ConanFile):
 
     def requirements(self):
         self.requires(self.tested_reference_str)
-        self.requires("opengl/system")
 
     def build(self):
         cmake = CMake(self)