+++ recipes/libselinux/all/conanfile.py
@@ -2,6 +2,7 @@ import os
 
 from conan import ConanFile
 from conan.errors import ConanInvalidConfiguration
+from conan.tools.build import cross_building
 from conan.tools.env import VirtualBuildEnv
 from conan.tools.files import apply_conandata_patches, chdir, copy, export_conandata_patches, get, rename, replace_in_file, save
 from conan.tools.gnu import Autotools, AutotoolsToolchain, PkgConfigDeps
@@ -62,7 +63,7 @@ class LibSELinuxConan(ConanFile):
     def build_requirements(self):
         self.tool_requires("flex/2.6.4")
         if not self.conf.get("tools.gnu:pkg_config", default=False, check_type=str):
-            self.tool_requires("pkgconf/2.0.3")
+            self.tool_requires("pkgconf/2.2.0")
 
     def source(self):
         for download in self.conan_data["sources"][self.version]:
@@ -87,7 +88,10 @@ class LibSELinuxConan(ConanFile):
         sepol_lib_folder = os.path.join(self._sepol_source_folder, "src")
         tc.extra_ldflags.append(f"-L{sepol_lib_folder}")
         tc.make_args.append("USE_PCRE2=y")
-        tc.generate()
+        env = tc.environment()
+        if cross_building(self):
+            env.append_path("PKG_CONFIG_LIBDIR", self.generators_folder)
+        tc.generate(env=env)
 
     def build(self):
         apply_conandata_patches(self)